backend/requirements.txt
Django==5.0
djangorestframework==3.14.0
django-environ==0.11.2
django-extensions==3.2.3
Werkzeug==3.0.1
black==24.1.1
flake8==7.0.0
isort==5.13.2
pre-commit==3.6.0
drf_yasg==1.21.7